+++ common-merge-structure/src/test/java/community/redrover/merge/util/FileUtilsTest.java
@@ -5,15 +5,14 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.junit.jupiter.api.Test;
-import java.io.IOException;
 import java.io.UncheckedIOException;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.Objects;
 import static org.junit.jupiter.api.Assertions.*;
+import community.redrover.merge.testutils.TempFile;
 
 public class FileUtilsTest {
 
@@ -43,36 +42,6 @@ public class FileUtilsTest {
             int result = name.hashCode();
             result = 31 * result + version;
             return result;
-        }
-    }
-
-    @Getter
-    private static class TempFile implements AutoCloseable {
-        private final Path path;
-
-        TempFile(String prefix, String suffix) {
-            try {
-                this.path = Files.createTempFile(prefix, suffix);
-            } catch (IOException e) {
-                throw new UncheckedIOException("Failed to create temp file with prefix '%s' and suffix '%s'".formatted(prefix, suffix), e);
-            }
-        }
-
-        public void write(String content) {
-            try {
-                Files.writeString(path, content);
-            } catch (IOException e) {
-                throw new UncheckedIOException("Failed to write to temp file: " + path, e);
-            }
-        }
-
-        @Override
-        public void close() {
-            try {
-                Files.deleteIfExists(path);
-            } catch (IOException e) {
-                throw new UncheckedIOException("Failed to delete existing temp file: " + path, e);
-            }
         }
     }
