+++ merge-structure-cli/pom.xml
@@ -38,5 +38,13 @@
             <version>5.10.0</version>
             <scope>test</scope>
         </dependency>
+
+        <dependency>
+            <groupId>community.redrover.merge</groupId>
+            <artifactId>common-merge-structure</artifactId>
+            <version>${project.version}</version>
+            <classifier>tests</classifier>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 </project>
