+++ merge-structure-cli/src/test/java/community/redrover/merge/cli/CliUtilsTest.java
@@ -6,56 +6,47 @@ import community.redrover.merge.model.config.AppendStrategyConfig;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import java.io.IOException;
-import java.io.UncheckedIOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.List;
 import static org.junit.jupiter.api.Assertions.*;
+import community.redrover.merge.testutils.TempFile;
 
 public class CliUtilsTest {
 
     @Test
-    void testIsInvalidPath_null() {
+    void testIsInvalidPathNull() {
         assertTrue(CliUtils.isInvalidPath(null));
     }
 
     @Test
-    void testIsInvalidPath_blank() {
+    void testIsInvalidPathBlank() {
         assertTrue(CliUtils.isInvalidPath("   "));
     }
 
     @Test
-    void testIsInvalidPath_startsWithDash() {
+    void testIsInvalidPathStartsWithDash() {
         assertTrue(CliUtils.isInvalidPath("--source"));
     }
 
     @Test
-    void testIsInvalidPath_valid() {
+    void testIsInvalidPathValid() {
         assertFalse(CliUtils.isInvalidPath("valid/path.yaml"));
     }
 
     @Test
-    void testParseArgs_helpFlag_throwsCliException() {
+    void testParseArgsHelpFlagThrowsCliException() {
         String[] args = {"--help"};
 
-        CliException ex = assertThrows(
-                CliException.class,
-                () -> CliUtils.parseArgs("append", args)
-        );
+        CliException ex = assertThrows(CliException.class, () -> CliUtils.parseArgs("append", args));
 
         assertTrue(ex.shouldShowUsage());
         assertNull(ex.getMessage());
     }
 
     @Test
-    void testParseArgs_invalidArgs_throwsCliException() {
+    void testParseArgsInvalidArgsThrowsCliException() {
         String[] args = {"--unknown"};
 
-        CliException ex = assertThrows(
-                CliException.class,
-                () -> CliUtils.parseArgs("append", args)
-        );
+        CliException ex = assertThrows(CliException.class, () -> CliUtils.parseArgs("append", args));
 
         assertTrue(ex.shouldShowUsage());
         assertTrue(ex.getMessage().startsWith("Error:"));
@@ -63,7 +54,7 @@ public class CliUtilsTest {
 
     @ParameterizedTest
     @ValueSource(strings = { "json", "yaml" })
-    void testLoadConfigOrUseArgs_withFallback(String format) {
+    void testLoadConfigOrUseArgsWithFallback(String format) {
         StrategyArgs mockArgs = new StrategyArgs();
         mockArgs.source      = "src." + format;
         mockArgs.destination = "dst." + format;
@@ -88,42 +79,9 @@ public class CliUtilsTest {
         assertEquals(expected, actual);
     }
 
-    static class TempFile implements AutoCloseable {
-        private final Path path;
-
-        TempFile(String prefix, String suffix) {
-            try {
-                this.path = Files.createTempFile(prefix, suffix);
-            } catch (IOException e) {
-                throw new UncheckedIOException("Failed to create temp file", e);
-            }
-        }
-
-        public Path getPath() {
-            return path;
-        }
-
-        public void write(String content) {
-            try {
-                Files.writeString(path, content);
-            } catch (IOException e) {
-                throw new UncheckedIOException("Failed to write to temp file: " + path, e);
-            }
-        }
-
-        @Override
-        public void close() {
-            try {
-                Files.deleteIfExists(path);
-            } catch (IOException e) {
-                throw new UncheckedIOException("Failed to delete temp file: " + path, e);
-            }
-        }
-    }
-
     @ParameterizedTest
     @ValueSource(strings = { "json", "yaml" })
-    void testLoadConfigOrUseArgs_withConfig(String format) {
+    void testLoadConfigOrUseArgsWithConfig(String format) {
         try (TempFile configFile = new TempFile("config", "." + format)) {
             if ("json".equals(format)) {
                 configFile.write("""
@@ -146,10 +104,7 @@ public class CliUtilsTest {
             StrategyArgs strategyArgs = new StrategyArgs();
             strategyArgs.config = configFile.getPath().toString();
 
-            ParsedStrategy parsed = new ParsedStrategy(
-                    strategyArgs,
-                    JCommander.newBuilder().addObject(strategyArgs).build()
-            );
+            ParsedStrategy parsed = new ParsedStrategy(strategyArgs, JCommander.newBuilder().addObject(strategyArgs).build());
 
             AppendStrategyConfig config = CliUtils.loadConfigOrUseArgs(
                     parsed,
@@ -174,19 +129,15 @@ public class CliUtilsTest {
         }
     }
 
-
     @ParameterizedTest
     @ValueSource(strings = { "json", "yaml" })
-    void testLoadConfigOrUseArgs_withArgsOnly(String format) {
+    void testLoadConfigOrUseArgsWithArgsOnly(String format) {
         StrategyArgs strategyArgs = new StrategyArgs();
         strategyArgs.source      = "src." + format;
         strategyArgs.destination = "dest." + format;
         strategyArgs.result      = "out." + format;
 
-        ParsedStrategy parsed = new ParsedStrategy(
-                strategyArgs,
-                JCommander.newBuilder().addObject(strategyArgs).build()
-        );
+        ParsedStrategy parsed = new ParsedStrategy(strategyArgs, JCommander.newBuilder().addObject(strategyArgs).build());
 
         AppendStrategyConfig config = CliUtils.loadConfigOrUseArgs(
                 parsed,
